+++ orchagent/routeorch.cpp
@@ -1,7 +1,7 @@
 #include <assert.h>
 #include "routeorch.h"
 #include "logger.h"
-#include "sainet.h"
+#include "swssnet.h"
 
 extern sai_next_hop_group_api_t*    sai_next_hop_group_api;
 extern sai_route_api_t*             sai_route_api;
