+++ orchagent/tunneldecaporch.cpp
@@ -1,7 +1,7 @@
 #include <string.h>
 #include "tunneldecaporch.h"
 #include "logger.h"
-#include "sainet.h"
+#include "swssnet.h"
 
 extern sai_tunnel_api_t* sai_tunnel_api;
 extern sai_router_interface_api_t* sai_router_intfs_api;
